Wait for IP Addresses cleared before transition to StoppedState.
To fix the race between unregistering NetworkAgent and stopping
IpClient, add an extra ClearAddressesOnStopState to clear addresses
and wait to see the RTM_DELADDR netlink messages. Set a configurable
timeout to wait for the netlink events, because netlink events will
get lost if ENOBUFS happens, then RTM_DELADDR will never arrive, it
results in never exiting ClearAddressesOnStopState.

Actually ClearingIpAddressesState is still required to cleanup the
leftover addresses. Addresses might be added after StoppedState due
to certain reason, then we still rely on ClearingIpAddressesState to
clear addresses before transtion to RunningState or PreconnectingState.
But in most cases the LinkProperties has been reset before transtion
to ClearingIpAddressesState, then it goes to the next state directly.

Consider to de-duplicate the states which do the same things, rename
ClearingIpAddressesState to ClearAddressesState for short and process
the same logic such as checking if it's ready to proceed to the next
state or should clear leftover addresses in advance.

ClearAddressesOnStartState is kind of simple refactor of the original
ClearingIpAddressesState by extending it, the behabvior should keep
the same as before. For ClearAddressesOnStopState, that's a new state,
use an experiment flag to control the new code path.

Bug: 240616892
Test: atest NetworkStackIntegrationTests
Test: atest NetworkStackTests
Change-Id: I0442fd1a998ee63ccf28689183a9a467ff919241