Fix presubmit crashes when signature tests can't find excludes.
Currently, when tradefed (e.g., presubmit) sees one or more test
failures, it will retry the failing tests. Specifically,
BaseRetryDecision lists the test that passed, and passes them to
AndroidJUnitTest, which writes them to an exclude file in
/data/local/tmp/ajur/excludes.txt and re-runs the module. This
results in re-running only the tests that failed.

This does not work for tests signed with the networkstack key
such as NetworkStackIntegrationTests, because the network_stack
selinux context does not have access to /data/local/tmp:

08-16 14:54:08.971  1073  6580  6580 W ntegrationtests: type=1400 audit(0.0:212): avc: denied { search } for name="tmp" dev="dm-49" ino=97 scontext=u:r:network_stack:s0 tcontext=u:object_r:shell_data_file:s0 tclass=dir permissive=0

This results in the module failing with a confusing error:

08-16 14:54:09.023  1073  6580  6580 E AndroidRuntime: Process: com.android.server.networkstack.integrationtests, PID: 6580
08-16 14:54:09.023  1073  6580  6580 E AndroidRuntime: java.lang.RuntimeException: Exception thrown in onCreate() of ComponentInfo{com.android.server.networkstack.integrationtests/androidx.test.runner.AndroidJUnitRunner}: java.lang.IllegalArgumentException: testfile not found: /data/local/tmp/ajur/excludes.txt
...
08-16 14:54:09.023  1073  6580  6580 E AndroidRuntime: Caused by: java.io.FileNotFoundException: /data/local/tmp/ajur/excludes.txt: open failed: EACCES (Permission denied)

Fix this by using test-filter-dir to specify the directory in
which to put the exclude files.

In order to do so, run the test as root, because without root
there appears to be no directory that shell can write to and
network_stack can read.

Test: test-only change
Change-Id: Id11c50c91d4d7396a624759edadaee8892f6312f