Increase valid dns time threshold for stall evaluation
The threshold is used to filter out those DNS queries that are
outdated and should not be considered in data stall evaluation.
The failed test will set mock DNS responses first and then
verify the isDataStall() result caused by those DNS responses.
This flow is expected to be handled soon, so 500ms is general
enough in most cases. But it may sometimes take more time and
fail the test. Thus, increase the timeout to 5000ms to ensure
the first expected DNS result generated within the threshold.

The threshold is used to verify the DNS result time only.
It does not increase the time that tests may take.

Fix: 242833380
Test: atest NetworkStackTests
Change-Id: I2edac60ff885f2f6c16c84fc264948f37db630fe