Only read test URLs on NetworkMonitor start
Instead of re-reading the test URLs every time there is a configuration
change, only read them once on NetworkMonitor start.

Tests currently assume that NetworkMonitor will not use the test URLs
unless the network was reconnected. This avoids breaking that assumption
in cases where the device has a configuration change.

Bug: 242495002
Test: atest NetworkStackTests CtsNetTestCases
Change-Id: I16863b654999fd549b8dc8ef2ad13d8f288b3fb8