Save ip_client_parse_netlink_events_version flag value at initialization.
From the public experiment we found that provisioning failure rate got
increased, that's caused by the potential switch from the old code path to new code path in the middle phase of connection.

For example:
- flag is off, refactor feature isn't enabled;
- doesn't register NetlinkMonitor and depend on netd callbacks to
  receive netlink updates;
- connected to wifi;
- push a new flag value to turn it on, the refactor new code path
  takes effect immediately;
- all upcoming netlink messages will get lost due to NetlinkMonitor
  isn't registered at initialization;
- IP provisioning failure happens after timeout(18s)

To prevent from reading the flag value and potential code switch in the
middle of connection, always save the flag value at initialization and
use that value until disconnection.

Bug: 246490949
Test: atest NetworkStackIntegrationTests
Change-Id: I0c83151896482dd573eb23da576d17710ddd50b2