Reduce the receive buffer size for the netlink event
The individual messages should not be excessively large(8-16kB)
because trying to get the kernel to return everything in one
big buffer is inefficient as it forces the kernel to allocate
large chunks of linearly physically contiguous memory. This
initial size for the first message is also limited in kernel.
TcpSocketTracker has to receive the message multiple times
anyway if there are more than 10 sockets. Reduce the receive
buffer size to have a better memory allocation.

Bug: 245696985
Test: atest NetworkStackTests
Test: manually test to check if all the socket information could
      retrieve successfully
Change-Id: I7563bdf5a610487fcc1b288523c15de369a21032