Fix data stall false alarm caused by TCP signal
Data stall detection with TCP signal refers to packets lost and
packets retransmission as a TCP fail rate to suspect a stall.
When device is idle, the retransmits may not be sent when the
network is restricted by doze mode. However, the counter of
retransmits will still increase since TCP stack will do the
statistic first then restricted later in the kernel. Thus,
stop check TCP signal for data stall detection if device is
idle to prevent this kind of false alarm.

Bug: 245484030
Test: manually test with wifi without internet to check if
      data stall detection works
Test: manually test with device idle and check the behavior of
      data stall detection
Test: atest NetworkStackTests
Change-Id: I774fe27699241d0c0e9285c045a7407076e06262