Add log to check the specific running NetworkStack module version.
This is to debug whether the test infra builds an older NetworkStack
version but runs the latest NetworkStackRootTests. If that's the case,
the production code will mismatch the test code, which results in the
broken test, such as we've fixed the type of vendor-specific IE type
in NetworkStack as well as the test code, but if test code is the latest
one but NetworkStack module is an older version which doesn't include
the fix, then test cases related to customized DHCP option will fail.

Bug: 256088652
Test: atest NetworkStackRootTests
Change-Id: Iaff4b700bb634a8259241362dd4ef468080ecacc