Remove DISMISS_PORTAL_IN_VALIDATED_NETWORK flag
The flag was introduced to allow toggling NetworkMonitor-based
validation in CaptivePortalLoginActivity, but was never pushed,
and the default (enabled on R+) has always been used instead.

Remove the flag as pushing it at this point would not be worth
the effort, considering that only Q devices do not use the
updated behavior. Rolling out the feature on Q would be risky
and have low value, and the feature has been in use for long
enough on R+ so that an emergency disable will not be necessary.
Also, the corresponding design was removed from CaptivePortalLogin.
As the comment in NetworkMonitor, this design shold be aligned
with CaptivePortalLogin design.

Bug: 259336759
Test: atest NetworkStackTests
Change-Id: I23b4aab60e6a02ad0ebc00d98aa9b72a4af93e10