Avoid to reset mocked IIpClientCallbacks in performDualStackProvisioning.
Notice that `when(mCb.getInterfaceVersion).thenReturn(aidl version)`
doesn't work any more after we reset mocked IIpClientCallbacks, we
leverage the specific return NetworkStack aidl version to check which
callback was invoked in the remaining test, but it always returns 0
after reset. However, reset mCb invocations is required in some test
cases such as testNoFdLeaks, which does loop to start dual-stack
provisioning to see if fd leak happens. Use clearInvocations()
instead of reset().

The goal of the "reset" was to allow verifying successive callbacks
without specifying "times(X)" every time if it was already called
previously. However it also clears setup of mocked methods, which is
an unintended side-effect. "clearInvocations" is the right method to
use to allow verifying successive invocations, not "reset".

Bug: 249550357
Test: atest NetworkStackIntegrationTests
Change-Id: I7990df03c93cc2ee0eaa33fc2e0e3e746b793254