Read the actual nud probe number from resource for NetworkStack root test.
So far the expected nud solicit num used in the root test is the default
config value for nud solicit num, however, the value can be overlaid via
RRO. For example, TV branch increases the overlay value to 15 for a long
duration of nud probe process. Therefore, we should read the actual
overlay config from the resource as the expected value instead of the
default value.

Bug: 265201014
Test: atest NetworkStackRootTests
Change-Id: Ia2f2c2d9a29aa77d61ff0a295c2b9e44a4b485c8