Make DnsUtils#generateFilter IP-agnostic.
Currently generateFilter can generate an IPv4-only filter or an
IPv6-only filter. This is not useful because a real APF filter
will need to work on both IPv4 and IPv6 packets.

Get the IP version from the packet being processed, so the same
filter will work for both IPv4 and IPv6 packets.

Also fix the method javadoc and slightly simplify the test.

Bug: 263429226
Test: passes existing tests
Change-Id: Ia8bf8072be1f258843d3fdd173f4b52b485abcce