Skip test if the tethering module is not new enough
For networkstack tests, it is not guaranteed that the tethering
module will be updated at the same time. If the tethering module
is not new enough, it may not contain the necessary abilities to
run these tests. For example, The tests depends on test network
stats being counted, which can only be achieved when they are
marked as TYPE_TEST.
If the tethering module does not support TYPE_TEST stats, then
these tests will need to be skipped.

Test: atest NetworkStackCoverageTests:android.net.NetworkStatsIntegrationTest
      on S device with newest NetworkStack module installed
Bug: 259632210
Change-Id: I640a5702ffb4a3778ab430713f35707da7ca86c8