Bump NetworkStack to target SDK 34
Bump NetworkStack to become NetworkStackNext, and create api35 shims
that are only included in the new NetworkStackNext.

This is a no-op on devices currently in the field as they use SDK 33
and there are no behavior changes affecting NetworkStack in 33

NetworkStackNext was also used until now by development API 34 builds,
so they will see no change either.

Ignore-AOSP-First: topic can't go to AOSP as-is, will cherrypick
Test: TH
Bug: 280250560
Change-Id: I8264dba7cddd515aade669fda869013b41a1fa7f