Add AIDL method to update the APF capabilities in IIpClient.aidl
The problem was that WiFi secondary interface doesn't support Apf at
provisioning as intended in STA+STA case, but primary interface does
support Apf. Due to the loss of apf capabilities settings on the
secondary interface Apf will not be enabled neither when wifi interface
switches from secondary to primary, but it's supposed to support Apf.
Add a new AIDL method in IIpClient.aidl, that allows WiFi to update the
Apf capabilities settings even after the provisioning completes.

Bug: 274090839
Test: atest NetworkStackTests
Change-Id: I9f19b622579177533080209c7024d17415065da8