Fix the flaky test: testIPv6LinkLocalOnly.
Ideally there should be only one route(fe80::/64 -> :: iface mtu 0)
in the LinkProperties when device is in the IPv6 link-local only mode,
but the multicast route(ff00::/8 -> :: iface mtu 0) may appear on some
old platforms where the kernel is still notifying the userspace the
multicast route. Therefore, we cannot assert that size of routes in
LinkProperties is more than one, but other properties such as DNS or
IPv6 default route or global IPv6 address should never appear in the
IPv6 link-local only mode.

Bug: 272302069
Test: atest NetworkStackRootTests
Change-Id: Id59d400d4a1ef0a309e38c94c09ff450704d74ff