Use a dedicated class to store stats
The previous test used a pair to get TX/RX bytes. This is being
changed to use a dedicated class that can also accommodate
packet counts. This will be useful for subsequent changes that
need to verify packet counts.

This is a no-op refactoring, which means that it does not change
the behavior of the test. However, it does make the test more
flexible and easier to maintain.

Test: atest NetworkStatsIntegrationTest
Bug: 276296921
Change-Id: I0bd5419b5921a968341221e768a0927e4ce920f8