Remove the exception thrown for adding a v6 keepalive data
System will translate v4 NAT-T keepalive packets to v6 so that
the keepalive packet could be sent to network. This makes v6 is
possible so remove the constraint.

Apf filter with a v6 NAT-T keepalive does not work with only
this change. This commit only remove the exception thrown
because of v6 address, but no apf filter will be updated.

Bug: 196453719
Test: m ; atest
Change-Id: Ia5bc7b93345db3acdad644424a58e229ab82d450