Support list of IaPrefixOption in PrefixDelegation class.
That allows parsing multiple IA prefix options included in one IA_PD
option, and send the list of IA prefix options in DHCPv6 request, renew
and rebind messages.

This CL supports a list of IA Prefix option in PrefixDelegation class
as the first step, but parsing the list of IA Prefix option isn't
supported yet, thus always read the first element of the list which
should be no-op change.

Bug: 260934173
Test: m
Change-Id: If7b477e1a3fe6bd4e7aae988c59ce2056b6098b6