Add null check for mPrefixDelegation in clearIpv6PrefixDelegationAddresses.
Ideally mPrefixDelegation should not be null when IpClient receives the
DHCPv6-PD failure messages (e.g. prefix expires or renewal prefix does
not match the request one), these failure should only happen after we
get a success PD. Add a null check in this function and Log.wtf to see
if we can hit this error in practice.

Bug: 290885706
Test: atest NetworkStackIntegrationTests
Change-Id: Ibfd2240d4cc3120c8e5de0cdf69a786a9b009118