Only use DhcpResults fromStableParcelable in tests
DhcpResults is now an internal class only used in NetworkStack, so it
does not need to be built from DhcpResultsParcelable (it is only
converted to a DhcpResultsParcelable to be delivered to Wifi).

Move fromStableParcelable to the tests as this is its only remaining
usage, and document that DhcpResultsParcelable should not receive any
new field there, and does not necessarily contain all information
necessary to build a DhcpResults.

Bug: 189865016
Test: atest
Change-Id: I4f4f3597276e32f85ef3925715a4fdc657cb538e