dhcpv6: schedule lease timers according to rfc8415
RFC8415 describes special mechanisms if t1 and / or t2 are 0, or t1 is
equal to t2 equals the valid lifetime. For the former, the client is
free to pick appropriate values for t1 / t2. For the latter, the server
indicates that the IA cannot be renewed, so there is no point in trying.

This change follows the recommended t1 / t2 deduction for servers (0.5
and 0.8 times the preferred lifetime). Special care must be taken so
timers stay in the right order; if timers happen to coincide, the first
one is cancelled.

Test: TH
Change-Id: I0f1f20e08f0618874213c2540593d95d9dd7a020