Move some DeviceConfig methods to common superclass.
This will simplify the code by using the same codepath in both
the root test and the signature test.

This CL does not actually change IpClientSignatureTest to use
DeviceConfig, but it moves the methods to the superclass so they
can be used in an upcoming CL to test code that uses
isNetworkStackFeatureNotChickenedOut.

The migration can be done in an upcoming CL.

Test: existing tests pass
Change-Id: I4c958fb66a45e8fc86a9fa276fec7a28a4362367