Update RA lifetimes when a new RA is received
This change instantiates a new RA object every time an RA is received,
which will cause result in lifetime sections being updated. The previous
code did a mere byte comparison of all MATCH sections, but did not
update any of its LIFETIME sections.

We may want to consider adding a Ra#inheritFromMatch(Ra oldRa) method
that handles moving data to the new Ra instance, but for now it is just
the seenCount (which may not be useful anyways and we could consider
deleting it).

Test: atest ApfTest
Bug: 299655294
Change-Id: Ie7aa49360642c8cb8f4ae2b468d96e9517c8eff1