Revert "Pass minimum acceptable RA lifetime to APF on creation."
This reverts commit 4aca303e00eaa4013d71972fc7009b843e8cb326.
Reason for revert: We want to create a change that can easily be
cherry-picked to the EoS module release branch (currently based on
tm(!)-mainline-prod). For now, it is easiest to revert these changes and
create a new, mostly self-contained change. Once we have that, the
original commits can be rebased on top.

Test: TH
Change-Id: Ib6783a8c6a2b513e8b98540598ef7eddd44aba32