Revert "Add unit test for accept_ra_min_lft sysctl used in IpClient."
This reverts commit 178079cd05f556680120e428aef32522b4865f43.
Reason for revert: We want to create a change that can easily be
cherry-picked to the EoS module release branch (currently based on
tm(!)-mainline-prod). For now, it is easiest to revert these changes and
create a new, mostly self-contained change. Once we have that, the
original commits can be rebased on top.

Test: TH
Change-Id: Ice59d5af4e79c4e9502d9b0769b3d95dc12c3bc5