Revert "Set accept_ra_min_lft systctl when starting IPv6 stack."
This reverts commit 3a68300d2663d8d7ebdcc322419bfe9b13094e33.
Reason for revert: We want to create a change that can easily be
cherry-picked to the EoS module release branch (currently based on
tm(!)-mainline-prod). For now, it is easiest to revert these changes and
create a new, mostly self-contained change. Once we have that, the
original commits can be rebased on top.

Test: TH
Change-Id: I154de34a6b3e4a8a6a582a679086461e9d8c18ba