Add additional asserts inside PacketSection constructor
Adds additional asserts inside the PacketSection constructor and throws
an IllegalArgumentException when any of them fail. This exception is
eventually caught when a new Ra is instantiated (which is an error case
that is already being handled).

This allows removing existing asserts from a few places in the code.

We should consider splitting PacketSection into two subclasses:
MatchSection and LifetimeSection.

This change also reverses the switch / case sorting order inside
generateFilterLocked() to ascending.

Test: atest ApfFilter
Change-Id: I059f3250e1ac4b870ac3052870b9a6710fd1f462