Add TokenBucket to rate limit filter installation
This change adds a token bucket to rate limit APF filter installation,
since basing a rate limit off the last filter installation time does not
work well for networks with 2 independent routers sending unsynchronized
RAs.

We do not expect to ever hit this limit (as APF filtering does not work
very well if filters are out-of-sync), but it is there to protect from
poorly configured networks that might be invoke the filter installation
path much more frequently than we would like to (it is assumed that
installing a new APF filter via HAL is an expensive operation).

Test: atest ApfTest
Change-Id: I1dc8d953783830e5371bbec3df8bfb951e2af57b