Use the ratio of total TCP retransmits as detection signal
The original fail rate is calculated from tcpi_lost and
tcpi_retransmits. The calculation uses the difference between
2 snapshots of the TCP socket statistic. The tcpi_retransmits
field is the number of recurring retransmission timeout of
first sequenced that is not acknowledged yet and will increase
once per retrasmit of a single packet. The tcpi_lost is the
current lost packets beyond the acked sequence and will increase
by one for packets that are not marked as TCPCB_LOST and
TCPCB_SACKED_ACKED. These two numbers are not of the same
dimension. This will cause the fail rate result over 100%.

In this case, tcpi_total_retrans is a better match as it's the
cumulative retransmitted packets. The retransmitted packets
should be a good enough signal because the network should be
in a bad shape if device keep sending retransmission without
acks. Thus, replace the indicator to use tcpi_total_retrans as
the nominator of the fail rate, and remove the unused lost
and retransmits fields in TcpInfo to reduce structure size.

Bug: 265107451
Test: atest NetworkStacktTests
Change-Id: Ib3ea3681fa5cbfaaef653ea53d12a7c9257db124