Rename PacketRetransmittingState to MessageExchangeState
rfc8415 refers to packet retransmissions as message exchange. So
renaming PacketRetransmittingState to MessageExchangeState gets us a bit
closer to the language used in the rfc.

I realize that PacketRetransmittingState was adopted from our DHCPv4
implementation; however, considering the complexity of DHCPv6, it is
unlikely that we will ever be able to merge parts of the two
implementations (as had previously been the plan). Doing so would only
needlessly complicate the code and increase the potential (negative)
impact of any change.

Bug: 260934173
Test: mm
Change-Id: Iaf39e9e899753c38a2ed3be2a0bb4a7ecdece4bb