network_stack_utils_jni.cpp - switch to classic bpf macros - part 5
This is a subtle change, we go from reading the ethertype
from the packet directly to reading it from skb->protocol metadata
field.  This means it should now work even for ethernet
header less packets (ie. rawip), or packets with extra
ethernet vlan tags, but it also means we're depending
on the kernel setting up skb->protocol correctly.

Test: TreeHugger, atest NetworkStackUtilsIntegrationTest
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: I153682a91222c3f2a0912b6fd22a700fa7999840