Stop purging RAs when they are expired
The previous logic was flawed as it purged RAs when they are expired,
not when their filter expired. In addition, there is not really a good
reason to remove RAs from mRas when their filters are expired.

Filter generation already skips RAs that would produce a 0-lifetime
filter. Similarly, this change now skips RA comparison in matches if the
tracked RA's filter has expired (similar to how an APF program
operates).

RAs are tracked in most-recently-seen order. When the list fills up, the
least-recently-seen RA is dropped from it. Without purging epired RAs
from the list, it is slightly more likely that an RA that would produce
an actual filter is dropped off that list. We might want to address this
in a follow up change.

Another side effect of this change is that mRas.size() is logged in
ApfProgramEvent; however, this is not a particularly useful piece of
information to log.

In addition, dump() will include expired RAs, which might be interesting
debug information.

Test: TH
Change-Id: I4592fb7d060849d8aaaa9fdc51c43cf00ee57631