Change MRT parameter type to IntSupplier in MessageExchangeState.
MRT parameter may change with the SOL_MAX_RT option value in
SolicitState, the default value is 3600s but server can replace
it by responding to Solicit with an valid SOL_MAX_RT option. Then
the MRT parameter used for Solicit message retransmission may change.
Using IntSupplier is easier to implement that and guarantee with final
member.

Also change the message retransmission contants type to int, which is
enough to represent the contants value.

Bug: 260934173
Test: TH
Change-Id: I455234773c2ba34818663fdfd1b365a6bf570c5b