Change the expected interactions
The TCP polling timer is set to 0 and NM in the test case
is expected to stay in the ValidatedState. This means the TCP
data stall may happen more than once. Test case expected only
one interaction is not realistic. Change it to verify at least
one interaction to solve the flaky test.

Also, update the timer to 0 to prevent possible busy loop.

Fix: 308920709
Test: atest NetworkStackTests
Change-Id: I42e384a3270ffc7b5010d6faef1595f5475d7989