Fix flaky because of multi-threads access
The testNotifyNetwork_WithforceReevaluation was flaky because
the test tried to mock the HttpConnection object for http probe
but NetworkMonitor still attempted sending the http probe. Both
threads(probing thread and test thread) access the same mock
object and result in CannotStubVoidMethodWithReturnValue
exception.

This purpose for this test does not have to mock the http probe
object again, so update the expected result to stop mocking
http probe again to prevent multi-threads access issue.

Fix: 311797297
Test: atest NetworkStackTests
Change-Id: Ib1eb3397f85b837e84c3dc40d8a8a6b0a2880a0f