Use PacketBuilderUtils in ApfFilterTest
Or a small part of it to figure out how useful it is.

Note that the RAs aren't entirely equivalent. scapy defaults all
lifetimes to infinite while PacketBuilderUtils tries to default to
commonly used rfc-recommended defaults.

In addition, the original version of ra1 also included a RA flags
extension option, which was omitted in the new version. ra2 originally
advertised an RIO with prefix /0 which was replaced with /48.

Test: TH
Flag: EXEMPT mainline
Bug: 1000001
Change-Id: I7c3d2bec390a7cc596c30c250871317912b2a0dd