Revert "Enable APF counter metrics on V+ devices with APFv4+"
This reverts commit ee6153e8a4d54c769affe4b4a27d0d8e11f6472f.

Reason for revert: Thinking through this carefully, I feel it might be better to wait for the dust to settle first. There are still a lot of devices that have a build exception for APF. Rolling this out too early might potentially trigger a WiFi HAL stuck on devices that have a build exception and still don't pass CTS yet. Rolling it out through mainline beta seems better.

Change-Id: Ib6c268bd25400aef5c72e2d2de3c5ec685bab6b8