Fix flaky testDhcp6Pd_multipleIaPrefixOptions.
Logcat analysis indicates that while an RS was successfully
sent from the tap interface and the test code responded with
an RA, this RA appears not to have been received by the tap
interface. Consequently, the tap interface is left without
a default IPv6 route, preventing the activation of the PD
code path.

A possible cause is that device's IPv6 stack is not fully enabled
when test code responds RS with a RA, therefore, the kernel missed
that RA and failed to get the default IPv6 route. Wait until the
IPv6 link-local address appears and then reply with RA, to see if
that helps.

Flag: EXEMPT test-only
Bug: 428036745
Test: atest NetworkStackIntegrationTests
Change-Id: I82b7e51f15e074af76b5656fee0e8c3bae1de500