Inline removeAddress
For readability. Also skip unnecessary isRegistrableAddress() check.
Note that this changes the value of shouldDispatchRegistration
and dispatchOnlyTimer; however, it is always safe to run
dispatchRegistration(). If there is no change, the function call results
in a noop. A follow up CL cleans this up a bit (by getting rid of
mLinkAddresses altogether).

Test: TH
Flag: EXEMPT mainline
Change-Id: I343bc459c8d73e6c27b4e08adce6ae85bb9f2984