Check for valid lifetime within 3s instead of 1%
The rfc9686 prescribed 1% threshold becomes smaller (i.e. approaches 0)
as the lifetime becomes shorter. It is also unclear how to handle
expired lifetimes which may occur in race scenarios / due to processing
delays.

Instead, check whether the new valid lifetime expires within 3s of the
original expiry. If so, ignore the update. This takes care of clock skew
and transmission delays and is much simpler to understand.

Test: TH
Flag: EXEMPT mainline
Change-Id: Ic32532d8827aa960d143dc612648c2f264daaa8d