Replace update with remove & add operation
This has the benefit that the RegistrationScheduler can now hold onto a
final LinkAddress. Previously, it was storing the Inet6Address
(originally the thought was that that could be made final to ensure it
can never be updated; however, this was not implemented correctly).
Updating an address means rescheduling the timer at an appropriate time
(either the original expiration time, or in 80% of the valid lifetime,
whichever occurs sooner) which can be handled directly inside
setLinkProperties.

Test: TH
Flag: EXEMPT mainline
Change-Id: I52d9d2a1a16f8688a63031331e9396e81c67a255