Store LinkAddress inside RegistrationScheduler
Previously, LinkAddress was moved out of RegistrationScheduler with the
idea that Inet6Address mAddress could be made final (however, that last
step was never done).

LinkAddress is now moved back inside RegistrationScheduler with the goal
of removing mLinkProperties in a follow up. Another follow up CL gets
rid of the update operation (and models it as a re-create) which allows
making mAddress final again.

Test: TH
Flag: EXEMPT mainline
Change-Id: Ia96db1f508d7018c9079e8b00998f78e41e65de5