Delete mLinkProperties
This change moves from comparing the new LinkProperties against the last
set of LinkProperties to comparing directly against what is stored
inside the RegistrationScheduler objects. The LinkAddress in that object
always contains the latest state of what was (or was attempted to be)
registered. This avoids a problem where a router might be sending RAs
with valid lifetime 1000 every 5 seconds, which (in the absence of APF
RA filtering) results in an address registration never being renewed
(because the lifetime changes less than 1%, but the mLinkProperties
object is still updated, so the next comparison also results in a < 1%
change).

Test: TH
Flag: EXEMPT mainline
Change-Id: Ie3640d4e4a8f477f0fc95681f8d3f300cc89367b