Ensure the registered valid / preferred lifetimes are never negative
Both getDeprecationTime() - now and getExpirationTime() - now can
realistically result in negative numbers (because a 0 lifetime RA was
received, or because at the time of renewal, the address is already
deprecated. Ensure that the registration packet never contains negative
numbers (which also results in jibberish when performing the downcast).

Test: TH
Flag: EXEMPT mainline
Change-Id: I5ba58cf139ebbce33bb7e663334df7c022c87f4d