Identify addresses by Inet6Address only
Ignore the prefix length. The same address with different prefix length
cannot (should not?) be present on the same interface at any time.

This also solves a problem that if the prefix length of an address
changes (does this ever happen?), the address is correctly put inside
the updated list, rather than the removed / added lists.

If an address should be identified by both address and prefix length,
then mTrackedAddresses should also be indexed by both -- but that makes
reply handling more difficult.

Test: TH
Flag: EXEMPT mainline
Change-Id: Iba3d33b20493d931ae2d9e3e03b05c2c54f759f1