Add APF generator convenience functions for move instruction
Generator function addMove(Register) moves a value from a source
register into a destination register (without modifying the source),
but when reading source code, it's not clear whether the specified
register is the source or destination, without consulting the API
docs.

So for clarity, add more explicitly named aliases for that function:
addMoveR0IntoR1 and addMoveR1IntoR0.

Also add tests for the move and swap instructions.

Bug: 408247593
Test: atest NetworkStackTests --test-filter ".*ApfGeneratorTest"
Flag: EXEMPT mainline
Change-Id: I73594fb4343dea99c3f570b3c5fb33a080ea22bf