Delay the client DUID initialization until addr tracker starts.
The client DUID was being initialized within the Dhcp6AddrRegTracker
constructor, which occurs in the IpClient constructor. In some cases,
this happens before the interface is ready, leading to a race condition
where interface parameters are null. This issue was observed during
runing the EthernetManagerTest CTS tests.

This CL delays fetching the client DUID until the Dhcp6AddrRegTracker
is started. This happens when the IpClient state machine enters
RunningState, a point where the interface is guaranteed to be ready for
use.

Flag: EXEMPT guarded by an legacy experiment flag
Bug: 293393743
Test: atest NetworkStackTests
Change-Id: Ifb90118b25459258e820da13d905ba64313828ac